Implemented Dashboard
This is where we show the list of polls
The user is be able to toggle between his/her answered and unanswered polls on the home page, which is located at the root.

The unanswered questions should be shown by default, and the name of the logged in user should be visible on the page.